Add the tally method, that will return useful info for the clients. Add some docs, specs and refactorings on coding style.
Important: replace defdelegate with def to #make_move on Hangman module.
the "tally" is intented to be used by the Client, so it makes sense that
the client interface is the only one that interacts with it.
This way, we removed the tally from the inner implementation of make_move
and moved it to the api boundaries.